Add a bunch of utility functions
I've moved a lot of information from the pseudo-hacked versions in their source files into
this module in order to clean it up a bit. I've also separated things via comments. Most notably,
our coordinate system will be independent of the screen's coordinate system in order to facilitate
the use of a camera. Our origin will start at the lower right-hand side of the screen, and the
camera is defined by a vector. Obviously the amount of screen that the camera sees depends on the
window size. The scale from screen-space to world-space is 100:1 (i.e. one hundred pixels correspond
to one unit in world space)