Fix input handling
Instead of adding an event to a list every time we register a keydown event, keep a list
of all of the currently accentuated events. Have it so that a keydown will maintain this
event in the list until we receive a key up, when we remove it. This way we get a continuous
event chain in the update loop without having to depend on keyboard repeating...