Add logic to update animated aabbs based on their coordinates.
In order to do this we needed to break apart the logic that iterates through animation
positions in our animation class. It's still a bit clunky because we specify these positions
in screen space sizes due to the dependency on StickyPy. It might be better to simply convert
them to world space whenever we load an animation.