Fix moving keyframes in keyframe editor.
This was done by changing what the 'editing' member variable holds in the keyframe editor. Previously
it only held references to keys that were being edited. These were not being propagated properly to
the actual data container because of a funky way that python deals with list references. In order to
work around this, the list of currenly edited keys contains not only the KeyFrame that's being edited, but
also the attribute to which the KeyFrame belongs. In this way, we can explicitly change the limbs after
we decide to drag the keys. This will also be useful for copy/paste functionality.