Fix problems with pasting keyframes.
In order to do this we also have to keep track of the limbs that are selected. When you copy/paste
a key, you must do it for every selected value. In order to do this our self.editing array also
needs to keep track of all of the limbs that are being selected as well.