Add a sphere particle domain, and check domain dimensions
Each domain responds to a dimension() function which returns the amount of
values that the domain generates. For example, spheres are three dimensional
objects because the values that they generate have an x, y, and z coordinate.
When we assign a domain to a particle property, we can make sure that the
dimensions correspond to the type of property it is...